requirements.txt
discord.py>=2.3.0
requests>=2.31.0
icalendar>=5.0.0
pytz>=2023.3
python-dotenv>=1.0.0
aiohttp
postgrest
supabase
